% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_heat_ER.R
\docType{data}
\name{CO_heat_ER_bygender}
\alias{CO_heat_ER_bygender}
\title{Subset of county-level age-adjusted heat-related ER visits between 2011 and 2022 by age category broken down by gender, as reported by the state of Colorado.}
\format{
A data frame with 240 rows and 7 columns.
\describe{
\item{county}{CO county which reported the datapoint}
\item{rate}{Age-adjusted rate of heat-related ER visits}
\item{lower95cl}{Lower 95\% confidence limit for the age-adjusted rate}
\item{upper95cl}{Upper 95\% confidence limit for the age-adjusted rate}
\item{visits}{Total number of heat-related ER visits}
\item{year}{Year between 2011 and 2022}
\item{gender}{Gender, with the option of 'female', 'male', and 'both genders'}
}
}
\source{
Created using data downloaded from the Colorado Environmental Public Health Tracking program website.
}
\usage{
CO_heat_ER_bygender
}
\description{
A dataset used in the DaSEH Short Course. This data is gathered by the Colorado Environmental Public Health Tracking program.
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
