% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_heat_ER.R
\docType{data}
\name{CO_heat_ER_long}
\alias{CO_heat_ER_long}
\title{Long-formatted dataset of age-adjusted heat-related ER visits between 2011 and 2022 for Boulder and Denver counties, as reported by the state of Colorado.}
\format{
A data frame with 24 rows and 3 columns.
\describe{
\item{county}{CO county, with the option of 'Boulder' or 'Denver'}
\item{rate}{Age-adjusted rate of heat-related ER visits}
\item{year}{Year between 2011 and 2022}
}
}
\source{
Created using data downloaded from the Colorado Environmental Public Health Tracking program website.
}
\usage{
CO_heat_ER_long
}
\description{
A dataset used in the DaSEH Short Course. This data is gathered by the Colorado Environmental Public Health Tracking program.
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
