% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_heat_ER.R
\docType{data}
\name{CO_heat_ER_wide}
\alias{CO_heat_ER_wide}
\title{Wide-formatted dataset of age-adjusted heat-related ER visits between 2011 and 2022 for Boulder and Denver counties, as reported by the state of Colorado.}
\format{
A data frame with 2 rows and 13 columns.
\describe{
\item{county}{CO county, with the option of 'Boulder' or 'Denver'}
\item{2011}{Age-adjusted rate for 2011}
\item{2012}{Age-adjusted rate for 2012}
\item{2013}{Age-adjusted rate for 2013}
\item{2014}{Age-adjusted rate for 2014}
\item{2015}{Age-adjusted rate for 2015}
\item{2016}{Age-adjusted rate for 2016}
\item{2017}{Age-adjusted rate for 2017}
\item{2018}{Age-adjusted rate for 2018}
\item{2019}{Age-adjusted rate for 2019}
\item{2020}{Age-adjusted rate for 2020}
\item{2021}{Age-adjusted rate for 2021}
\item{2022}{Age-adjusted rate for 2022}
}
}
\source{
Created using data downloaded from the Colorado Environmental Public Health Tracking program website.
}
\usage{
CO_heat_ER_wide
}
\description{
A dataset used in the DaSEH Short Course. This data is gathered by the Colorado Environmental Public Health Tracking program.
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
