% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_wastewater.R
\docType{data}
\name{covid_wastewater}
\alias{covid_wastewater}
\title{SARS-CoV-2 Wastewater Data}
\format{
A data frame with x rows and x columns.
\describe{
\item{reporting_jurisdiction}{}
\item{sample_location}{}
\item{key_plot_id}{}
\item{county_names}{}
\item{population_served}{}
\item{date_start}{}
\item{date_end}{}
\item{rna_pct_change_15d}{}
\item{pos_PCR_prop_15d}{}
\item{percentile}{}
\item{sampling_prior}{}
\item{first_sample_date}{}
}
}
\source{
Created using data downloaded from x.
}
\usage{
covid_wastewater
}
\description{
A dataset used in the DaSEH Short Course. More here..
}
\examples{
data(covid_wastewater)
}
\keyword{datasets}
