% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_CO.R
\docType{data}
\name{er_CO_statewide}
\alias{er_CO_statewide}
\title{Statewide age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. No breakdown by county, gender, or age category.}
\format{
A data frame with 12 rows and 5 columns.
\describe{
\item{rate}{Age-adjusted rate of heat-related ER visits}
\item{lower95cl}{Lower 95\% confidence limit for the age-adjusted rate}
\item{upper95cl}{Upper 95\% confidence limit for the age-adjusted rate}
\item{visits}{Total number of heat-related ER visits}
\item{year}{Year between 2011 and 2022}
}
}
\source{
Created using data downloaded from the Colorado Environmental Public Health Tracking program website
}
\usage{
er_CO_statewide
}
\description{
A dataset used in the DaSEH Short Course. This data is gathered by the Colorado Environmental Public Health Tracking program.
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
