% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_CO.R
\docType{data}
\name{er_visits_by_CO_county}
\alias{er_visits_by_CO_county}
\title{Colorado Heat-related Emergency Room Visits (Age Adjusted) Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 205 rows and 6 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
er_visits_by_CO_county
}
\description{
A dataset used in the DaSEH Short Course. Data is in "long" form. More here..
}
\examples{
data(er_visits_by_CO_county)
}
\keyword{datasets}
