% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haa5.R
\docType{data}
\name{haa5}
\alias{haa5}
\title{HAA5 Exposure for Washington State Public Water Systems Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 33 rows and 11 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
haa5
}
\description{
A dataset used in the DaSEH Short Course. More here..
}
\examples{
data(haa5)
}
\keyword{datasets}
