% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitrate.R
\docType{data}
\name{nitrate}
\alias{nitrate}
\title{Nitrate Exposure for Washington State Public Water Systems By Quarter Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 88 rows and 11 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
nitrate
}
\description{
A dataset used in the DaSEH Short Course. More here..
}
\examples{
data(nitrate)
}
\keyword{datasets}
