% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_co2_emissions.R
\docType{data}
\name{yearly_co2_emissions}
\alias{yearly_co2_emissions}
\title{Estimated yearly CO2 emissions in thousands of tonnes by country, 1751-2014}
\format{
A data frame with 192 rows and 265 columns.
\describe{
  \item{country}{Country name}
  \item{1751}{Emissions for the year 1751}
  \item{1752}{Emissions for the year 1752}
  \item{1753}{Emissions for the year 1753}
  \item{1754}{Emissions for the year 1754}
  \item{1755}{Emissions for the year 1755}
  \item{1756}{Emissions for the year 1756}
  \item{1757}{Emissions for the year 1757}
  \item{1758}{Emissions for the year 1758}
  \item{1759}{Emissions for the year 1759}
  \item{1760}{Emissions for the year 1760}
  \item{1761}{Emissions for the year 1761}
  \item{1762}{Emissions for the year 1762}
  \item{1763}{Emissions for the year 1763}
  \item{1764}{Emissions for the year 1764}
  \item{1765}{Emissions for the year 1765}
  \item{1766}{Emissions for the year 1766}
  \item{1767}{Emissions for the year 1767}
  \item{1768}{Emissions for the year 1768}
  \item{1769}{Emissions for the year 1769}
  \item{1770}{Emissions for the year 1770}
  \item{1771}{Emissions for the year 1771}
  \item{1772}{Emissions for the year 1772}
  \item{1773}{Emissions for the year 1773}
  \item{1774}{Emissions for the year 1774}
  \item{1775}{Emissions for the year 1775}
  \item{1776}{Emissions for the year 1776}
  \item{1777}{Emissions for the year 1777}
  \item{1778}{Emissions for the year 1778}
  \item{1779}{Emissions for the year 1779}
  \item{1780}{Emissions for the year 1780}
  \item{1781}{Emissions for the year 1781}
  \item{1782}{Emissions for the year 1782}
  \item{1783}{Emissions for the year 1783}
  \item{1784}{Emissions for the year 1784}
  \item{1785}{Emissions for the year 1785}
  \item{1786}{Emissions for the year 1786}
  \item{1787}{Emissions for the year 1787}
  \item{1788}{Emissions for the year 1788}
  \item{1789}{Emissions for the year 1789}
  \item{1790}{Emissions for the year 1790}
  \item{1791}{Emissions for the year 1791}
  \item{1792}{Emissions for the year 1792}
  \item{1793}{Emissions for the year 1793}
  \item{1794}{Emissions for the year 1794}
  \item{1795}{Emissions for the year 1795}
  \item{1796}{Emissions for the year 1796}
  \item{1797}{Emissions for the year 1797}
  \item{1798}{Emissions for the year 1798}
  \item{1799}{Emissions for the year 1799}
  \item{1800}{Emissions for the year 1800}
  \item{1801}{Emissions for the year 1801}
  \item{1802}{Emissions for the year 1802}
  \item{1803}{Emissions for the year 1803}
  \item{1804}{Emissions for the year 1804}
  \item{1805}{Emissions for the year 1805}
  \item{1806}{Emissions for the year 1806}
  \item{1807}{Emissions for the year 1807}
  \item{1808}{Emissions for the year 1808}
  \item{1809}{Emissions for the year 1809}
  \item{1810}{Emissions for the year 1810}
  \item{1811}{Emissions for the year 1811}
  \item{1812}{Emissions for the year 1812}
  \item{1813}{Emissions for the year 1813}
  \item{1814}{Emissions for the year 1814}
  \item{1815}{Emissions for the year 1815}
  \item{1816}{Emissions for the year 1816}
  \item{1817}{Emissions for the year 1817}
  \item{1818}{Emissions for the year 1818}
  \item{1819}{Emissions for the year 1819}
  \item{1820}{Emissions for the year 1820}
  \item{1821}{Emissions for the year 1821}
  \item{1822}{Emissions for the year 1822}
  \item{1823}{Emissions for the year 1823}
  \item{1824}{Emissions for the year 1824}
  \item{1825}{Emissions for the year 1825}
  \item{1826}{Emissions for the year 1826}
  \item{1827}{Emissions for the year 1827}
  \item{1828}{Emissions for the year 1828}
  \item{1829}{Emissions for the year 1829}
  \item{1830}{Emissions for the year 1830}
  \item{1831}{Emissions for the year 1831}
  \item{1832}{Emissions for the year 1832}
  \item{1833}{Emissions for the year 1833}
  \item{1834}{Emissions for the year 1834}
  \item{1835}{Emissions for the year 1835}
  \item{1836}{Emissions for the year 1836}
  \item{1837}{Emissions for the year 1837}
  \item{1838}{Emissions for the year 1838}
  \item{1839}{Emissions for the year 1839}
  \item{1840}{Emissions for the year 1840}
  \item{1841}{Emissions for the year 1841}
  \item{1842}{Emissions for the year 1842}
  \item{1843}{Emissions for the year 1843}
  \item{1844}{Emissions for the year 1844}
  \item{1845}{Emissions for the year 1845}
  \item{1846}{Emissions for the year 1846}
  \item{1847}{Emissions for the year 1847}
  \item{1848}{Emissions for the year 1848}
  \item{1849}{Emissions for the year 1849}
  \item{1850}{Emissions for the year 1850}
  \item{1851}{Emissions for the year 1851}
  \item{1852}{Emissions for the year 1852}
  \item{1853}{Emissions for the year 1853}
  \item{1854}{Emissions for the year 1854}
  \item{1855}{Emissions for the year 1855}
  \item{1856}{Emissions for the year 1856}
  \item{1857}{Emissions for the year 1857}
  \item{1858}{Emissions for the year 1858}
  \item{1859}{Emissions for the year 1859}
  \item{1860}{Emissions for the year 1860}
  \item{1861}{Emissions for the year 1861}
  \item{1862}{Emissions for the year 1862}
  \item{1863}{Emissions for the year 1863}
  \item{1864}{Emissions for the year 1864}
  \item{1865}{Emissions for the year 1865}
  \item{1866}{Emissions for the year 1866}
  \item{1867}{Emissions for the year 1867}
  \item{1868}{Emissions for the year 1868}
  \item{1869}{Emissions for the year 1869}
  \item{1870}{Emissions for the year 1870}
  \item{1871}{Emissions for the year 1871}
  \item{1872}{Emissions for the year 1872}
  \item{1873}{Emissions for the year 1873}
  \item{1874}{Emissions for the year 1874}
  \item{1875}{Emissions for the year 1875}
  \item{1876}{Emissions for the year 1876}
  \item{1877}{Emissions for the year 1877}
  \item{1878}{Emissions for the year 1878}
  \item{1879}{Emissions for the year 1879}
  \item{1880}{Emissions for the year 1880}
  \item{1881}{Emissions for the year 1881}
  \item{1882}{Emissions for the year 1882}
  \item{1883}{Emissions for the year 1883}
  \item{1884}{Emissions for the year 1884}
  \item{1885}{Emissions for the year 1885}
  \item{1886}{Emissions for the year 1886}
  \item{1887}{Emissions for the year 1887}
  \item{1888}{Emissions for the year 1888}
  \item{1889}{Emissions for the year 1889}
  \item{1890}{Emissions for the year 1890}
  \item{1891}{Emissions for the year 1891}
  \item{1892}{Emissions for the year 1892}
  \item{1893}{Emissions for the year 1893}
  \item{1894}{Emissions for the year 1894}
  \item{1895}{Emissions for the year 1895}
  \item{1896}{Emissions for the year 1896}
  \item{1897}{Emissions for the year 1897}
  \item{1898}{Emissions for the year 1898}
  \item{1899}{Emissions for the year 1899}
  \item{1900}{Emissions for the year 1900}
  \item{1901}{Emissions for the year 1901}
  \item{1902}{Emissions for the year 1902}
  \item{1903}{Emissions for the year 1903}
  \item{1904}{Emissions for the year 1904}
  \item{1905}{Emissions for the year 1905}
  \item{1906}{Emissions for the year 1906}
  \item{1907}{Emissions for the year 1907}
  \item{1908}{Emissions for the year 1908}
  \item{1909}{Emissions for the year 1909}
  \item{1910}{Emissions for the year 1910}
  \item{1911}{Emissions for the year 1911}
  \item{1912}{Emissions for the year 1912}
  \item{1913}{Emissions for the year 1913}
  \item{1914}{Emissions for the year 1914}
  \item{1915}{Emissions for the year 1915}
  \item{1916}{Emissions for the year 1916}
  \item{1917}{Emissions for the year 1917}
  \item{1918}{Emissions for the year 1918}
  \item{1919}{Emissions for the year 1919}
  \item{1920}{Emissions for the year 1920}
  \item{1921}{Emissions for the year 1921}
  \item{1922}{Emissions for the year 1922}
  \item{1923}{Emissions for the year 1923}
  \item{1924}{Emissions for the year 1924}
  \item{1925}{Emissions for the year 1925}
  \item{1926}{Emissions for the year 1926}
  \item{1927}{Emissions for the year 1927}
  \item{1928}{Emissions for the year 1928}
  \item{1929}{Emissions for the year 1929}
  \item{1930}{Emissions for the year 1930}
  \item{1931}{Emissions for the year 1931}
  \item{1932}{Emissions for the year 1932}
  \item{1933}{Emissions for the year 1933}
  \item{1934}{Emissions for the year 1934}
  \item{1935}{Emissions for the year 1935}
  \item{1936}{Emissions for the year 1936}
  \item{1937}{Emissions for the year 1937}
  \item{1938}{Emissions for the year 1938}
  \item{1939}{Emissions for the year 1939}
  \item{1940}{Emissions for the year 1940}
  \item{1941}{Emissions for the year 1941}
  \item{1942}{Emissions for the year 1942}
  \item{1943}{Emissions for the year 1943}
  \item{1944}{Emissions for the year 1944}
  \item{1945}{Emissions for the year 1945}
  \item{1946}{Emissions for the year 1946}
  \item{1947}{Emissions for the year 1947}
  \item{1948}{Emissions for the year 1948}
  \item{1949}{Emissions for the year 1949}
  \item{1950}{Emissions for the year 1950}
  \item{1951}{Emissions for the year 1951}
  \item{1952}{Emissions for the year 1952}
  \item{1953}{Emissions for the year 1953}
  \item{1954}{Emissions for the year 1954}
  \item{1955}{Emissions for the year 1955}
  \item{1956}{Emissions for the year 1956}
  \item{1957}{Emissions for the year 1957}
  \item{1958}{Emissions for the year 1958}
  \item{1959}{Emissions for the year 1959}
  \item{1960}{Emissions for the year 1960}
  \item{1961}{Emissions for the year 1961}
  \item{1962}{Emissions for the year 1962}
  \item{1963}{Emissions for the year 1963}
  \item{1964}{Emissions for the year 1964}
  \item{1965}{Emissions for the year 1965}
  \item{1966}{Emissions for the year 1966}
  \item{1967}{Emissions for the year 1967}
  \item{1968}{Emissions for the year 1968}
  \item{1969}{Emissions for the year 1969}
  \item{1970}{Emissions for the year 1970}
  \item{1971}{Emissions for the year 1971}
  \item{1972}{Emissions for the year 1972}
  \item{1973}{Emissions for the year 1973}
  \item{1974}{Emissions for the year 1974}
  \item{1975}{Emissions for the year 1975}
  \item{1976}{Emissions for the year 1976}
  \item{1977}{Emissions for the year 1977}
  \item{1978}{Emissions for the year 1978}
  \item{1979}{Emissions for the year 1979}
  \item{1980}{Emissions for the year 1980}
  \item{1981}{Emissions for the year 1981}
  \item{1982}{Emissions for the year 1982}
  \item{1983}{Emissions for the year 1983}
  \item{1984}{Emissions for the year 1984}
  \item{1985}{Emissions for the year 1985}
  \item{1986}{Emissions for the year 1986}
  \item{1987}{Emissions for the year 1987}
  \item{1988}{Emissions for the year 1988}
  \item{1989}{Emissions for the year 1989}
  \item{1990}{Emissions for the year 1990}
  \item{1991}{Emissions for the year 1991}
  \item{1992}{Emissions for the year 1992}
  \item{1993}{Emissions for the year 1993}
  \item{1994}{Emissions for the year 1994}
  \item{1995}{Emissions for the year 1995}
  \item{1996}{Emissions for the year 1996}
  \item{1997}{Emissions for the year 1997}
  \item{1998}{Emissions for the year 1998}
  \item{1999}{Emissions for the year 1999}
  \item{2000}{Emissions for the year 2000}
  \item{2001}{Emissions for the year 2001}
  \item{2002}{Emissions for the year 2002}
  \item{2003}{Emissions for the year 2003}
  \item{2004}{Emissions for the year 2004}
  \item{2005}{Emissions for the year 2005}
  \item{2006}{Emissions for the year 2006}
  \item{2007}{Emissions for the year 2007}
  \item{2008}{Emissions for the year 2008}
  \item{2009}{Emissions for the year 2009}
  \item{2010}{Emissions for the year 2010}
  \item{2011}{Emissions for the year 2011}
  \item{2012}{Emissions for the year 2012}
  \item{2013}{Emissions for the year 2013}
  \item{2014}{Emissions for the year 2014}
}
}
\source{
Created using data downloaded from \url{https://data.ess-dive.lbl.gov/portals/CDIAC}.
}
\usage{
yearly_co2_emissions
}
\description{
A dataset used in the DaSEH Short Course. Data gathered by the Carbon Dioxide Information Analysis Center (CDIAC).
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
