% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_heat_ER.R
\docType{data}
\name{CO_heat_ER_byage}
\alias{CO_heat_ER_byage}
\title{Colorado Heat-related Emergency Room Visits (Age Adjusted) by Age Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 216 rows and 7 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
CO_heat_ER_byage
}
\description{
A dataset used in the DaSEH Short Course. More here..
}
\examples{
data(CO_heat_ER_byage)
}
\keyword{datasets}
