% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_heat_ER.R
\docType{data}
\name{CO_heat_ER_long}
\alias{CO_heat_ER_long}
\title{Colorado Heat-related Emergency Room Visits (Age Adjusted) Data.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 24 rows and 3 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
CO_heat_ER_long
}
\description{
A dataset used in the DaSEH Short Course. This data is in long, or "tidy",
form. More here..
}
\examples{
data(CO_heat_ER_long)
}
\keyword{datasets}
