% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasehr-package.R
\docType{package}
\name{dasehr-package}
\alias{dasehr}
\alias{dasehr-package}
\title{`dasehr`: Load data and packages needed for DaSEH (Data Science for
Environmental Health)}
\description{
This package contains the datasets used in the Data Science for Environmental Health (DaSEH) short course taught by Ava M. Hoffman and Carrie Wright through the Data Science Lab at the Fred Hutchinson Cancer Center.
}
\details{
Here is a list of the datasets. They are already stored as objects when you
load the package.
* `calenviroscreen`: Description here
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ehumph/dasehr}
  \item Report bugs at \url{https://github.com/ehumph/dasehr/issues}
}

}
\keyword{internal}
