% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasehr-package.R
\docType{package}
\name{dasehr-package}
\alias{dasehr}
\alias{dasehr-package}
\title{`dasehr`: Load data and packages needed for DaSEH (Data Science for
Environmental Health)}
\description{
This package contains the datasets used in the Data Science for Environmental Health (DaSEH) short course taught by Ava M. Hoffman and Carrie Wright through the Data Science Lab at the Fred Hutchinson Cancer Center.
}
\details{
Here is a list of the datasets. They are already stored as objects when you
load the package.

* `calenviroscreen`: CalEnviroScreen4.0 data. This dataset was gathered by the California Office of Environmental Health Hazard Assessment. CalEnviroScreen ranks census tracts in California based on potential exposures to pollutants, adverse environmental conditions, socioeconomic factors and the prevalence of certain health conditions. Read more at https://calenviroscreen-oehha.hub.arcgis.com/

* `CO_heat_ER_byage`: Age-group breakdown of heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `CO_heat_ER_bygender`: Gender and county-level breakdown of age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `CO_heat_ER_wide`: Boulder and Denver counties' age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `CO_heat_ER`: Age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `covid_wastewater`: Covid levels in wastewater June 2020 - May 2024, as tracked by the National Wastewater Surveillance System. Read more at https://covid.cdc.gov/covid-data-tracker/#wastewater-surveillance

* `er_CO_statewide`: Age-adjusted statewide heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `er_visits_by_CO_county`: Age-adjusted county-level heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at https://coepht.colorado.gov/heat-related-illness

* `haa5`: Haloacetic acid exposure for populations on public water systems in the state of Washington for 1999-2020. This dataset was collected by the Washington Tracking Network. Read more at https://doh.wa.gov/data-and-statistical-reports/washington-tracking-network-wtn/drinking-water

* `nitrate`: Nitrate exposure for populations on public water systems in the state of Washington for 1999-2020, divided by quarter. This dataset was collected by the Washington Tracking Network. Read more at https://doh.wa.gov/data-and-statistical-reports/washington-tracking-network-wtn/drinking-water

* `yearly_co2_emissions`: Estimated yearly CO2 emissions by country, 1751-2014. Data borrowed from the Open Case Studies project https://www.opencasestudies.org/ocs-bp-co2-emissions/. Data originally gathered by the Carbon Dioxide Information Analysis Center (CDIAC). Read more at https://data.ess-dive.lbl.gov/portals/CDIAC
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhdsl/dasehr}
  \item Report bugs at \url{https://github.com/fhdsl/dasehr/issues}
}

}
\author{
\strong{Maintainer}: Elizabeth Humphries \email{ehumphri@fredhutch.org}

Authors:
\itemize{
  \item Ava Hoffman \email{avamariehoffman@gmail.com} (\href{https://orcid.org/0000-0002-1833-4397}{ORCID})
}

}
\keyword{internal}
