% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dasehr-package.R
\docType{package}
\name{dasehr-package}
\alias{dasehr}
\alias{dasehr-package}
\title{`dasehr`: Load data and packages needed for DaSEH (Data Science for
Environmental Health)}
\description{
This package contains the datasets used in the Data Science for Environmental Health (DaSEH) short course taught by Ava M. Hoffman and Carrie Wright through the Data Science Lab at the Fred Hutchinson Cancer Center.
}
\details{
Here is a list of the datasets. They are already stored as objects when you
load the package.

 * `calenviroscreen`: CalEnviroScreen4.0 data. This dataset was gathered by the California Office of Environmental Health Hazard Assessment. CalEnviroScreen ranks census tracts in California based on potential exposures to pollutants, adverse environmental conditions, socioeconomic factors and the prevalence of certain health conditions. Read more at \url{https://calenviroscreen-oehha.hub.arcgis.com/}.

 * `CO_heat_ER_byage`: Age-group breakdown of heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}

 * `CO_heat_ER_bygender`: Gender and county-level breakdown of age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}.

 * `CO_heat_ER_wide`: Boulder and Denver counties' age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}.

 * `CO_heat_ER`: Age-adjusted heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}.

 * `covid_wastewater`: SARS-CoV-2 levels in wastewater June 2020 - May 2024, as tracked by the National Wastewater Surveillance System. Read more at \url{https://data.cdc.gov/Public-Health-Surveillance/NWSS-Public-SARS-CoV-2-Wastewater-Metric-Data/2ew6-ywp6/about_data}.

 * `er_CO_statewide`: Age-adjusted statewide heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}.

 * `er_CO_county`: Age-adjusted county-level heat-related ER visits between 2011 and 2022, as reported by the state of Colorado. This data is gathered by the Colorado Environmental Public Health Tracking program. Read more at \url{https://coepht.colorado.gov/heat-related-illness}.

 * `haa5`: Haloacetic acid exposure for populations on public water systems in the state of Washington for 1999-2020. This dataset was collected by the Washington Tracking Network. Read more at \url{https://doh.wa.gov/data-and-statistical-reports/washington-tracking-network-wtn/drinking-water}.

 * `nitrate`: Nitrate exposure for populations on public water systems in the state of Washington for 1999-2020, divided by quarter. This dataset was collected by the Washington Tracking Network. Read more at \url{https://doh.wa.gov/data-and-statistical-reports/washington-tracking-network-wtn/drinking-water}.

 * `yearly_co2_emissions`: Estimated yearly CO2 emissions by country, 1751-2014. Data borrowed from the Open Case Studies project \url{https://www.opencasestudies.org/ocs-bp-co2-emissions/}. Data originally gathered by the Carbon Dioxide Information Analysis Center (CDIAC). Read more at \url{https://data.ess-dive.lbl.gov/portals/CDIAC}.
}
\seealso{
\code{\link{calenviroscreen}}

\code{\link{CO_heat_ER_byage}}

\code{\link{CO_heat_ER_bygender}}

\code{\link{CO_heat_ER_wide}}

\code{\link{CO_heat_ER}}

\code{\link{covid_wastewater}}

\code{\link{er_CO_statewide}}

\code{\link{er_CO_county}}

\code{\link{haa5}}

\code{\link{nitrate}}

\code{\link{yearly_co2_emissions}}
}
\author{
\strong{Maintainer}: Elizabeth Humphries \email{emarellahumphries@gmail.com}

Authors:
\itemize{
  \item Ava Hoffman \email{avamariehoffman@gmail.com} (\href{https://orcid.org/0000-0002-1833-4397}{ORCID})
}

}
\keyword{internal}
