% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/er_CO.R
\docType{data}
\name{er_CO_statewide}
\alias{er_CO_statewide}
\title{Colorado Statewide Heat-related Emergency Room Visits (Age Adjusted) Data.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12 rows and 5 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
er_CO_statewide
}
\description{
A dataset used in the DaSEH Short Course. Data is in "long" form. More here..
}
\examples{
data(er_CO_statewide)
}
\keyword{datasets}
