% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_co2_emissions.R
\docType{data}
\name{yearly_co2_emissions}
\alias{yearly_co2_emissions}
\title{Estimated yearly CO2 emissions in thousands of tonnes by country, 1751-2014}
\format{
A data frame with 192 rows and 265 columns.
\describe{
\item{country}{Unique country name}
}
}
\source{
Created using data downloaded from \url{https://data.ess-dive.lbl.gov/portals/CDIAC}.
}
\usage{
yearly_co2_emissions
}
\description{
A dataset used in the DaSEH Short Course. Data gathered by the Carbon Dioxide Information Analysis Center (CDIAC).
}
\seealso{
\code{\link{dasehr-package}}
}
\keyword{datasets}
