% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yearly_co2_emissions.R
\docType{data}
\name{yearly_co2_emissions}
\alias{yearly_co2_emissions}
\title{Yearly CO2 Emissions in 1000s of Tonnes Data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 192 rows and 265 columns.
}
\source{
Created using data downloaded from x.
}
\usage{
yearly_co2_emissions
}
\description{
A dataset used in the DaSEH Short Course. More here..
}
\examples{
data(yearly_co2_emissions)
}
\keyword{datasets}
